Add base routing for date and fix respo
